[SDK] Add read-only access to the StateStore
